Fix date parsing to and from objects
In order to fix a bug where dates were not correctly translated back and
forth between string representations, this commit uses the propertyClass
and not the keyClass when checking for NSDate subclasses. This is
because we want to base the date translation on the property declaration
of a model and not the value from the dictionary (which is a string). In
addition, when translating back to dictionaries we set the dateString on
the dictionary we are returning and not the original object passed in.